refactor!: flatten namespaces — Utopia\Lock\Lock\* → Utopia\Lock\*
