+++ src/Semaphore.php
@@ -2,13 +2,12 @@
 
 declare(strict_types=1);
 
-namespace Utopia\Lock\Lock;
+namespace Utopia\Lock;
 
 use InvalidArgumentException;
 use Swoole\Coroutine;
 use Swoole\Coroutine\Channel;
 use Utopia\Lock\Exception\Contention;
-use Utopia\Lock\Lock;
 
 final class Semaphore implements Lock
 {
